Add 1007,1008 & 1009 attribute info to version info on linux, & 1009 on Windows. Add -version option to Mac VM (with similar info to linux). Fix compilation warning in sqWin32PluginSupport.c
git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@2530 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 2b670b32f9c4a633470b7b5f25954a759beb6508

+++ platforms/win32/vm/sqWin32PluginSupport.c
@@ -250,7 +250,7 @@ EXPORT(int) primitivePluginRequestFileHandle(void)
 #endif
     fileHandle = instantiateClassindexableSize(classByteArray(), fileRecordSize());
     fBrowserMode = false;
-    sqFileOpen(fileValueOf(fileHandle),(int)req->localName, strlen(req->localName), 0);
+    sqFileOpen(fileValueOf(fileHandle),req->localName,strlen(req->localName),0);
     fBrowserMode = true;
     if(failed()) return 0;
   }
